Refactor create solomachine client message builder
