+++ crates/relayer-solomachine/src/impls/chain/cosmos_components/mod.rs
@@ -1,4 +1,5 @@
 pub mod channel_handshake_message;
 pub mod connection_handshake_message;
+pub mod create_client_message;
 pub mod query_client_state;
 pub mod query_consensus_state;
